fetch inbox items from firestore
