+++ src/services/utils/index.ts
@@ -1,2 +1,3 @@
+export * from "./addMetadataToItemsBatch";
 export * from "./waitForCommonToBeLoaded";
 export * from "./waitForUserToBeLoaded";
