+++ src/shared/hooks/useCases/useInboxItems.ts
@@ -1,27 +1,13 @@
 import { useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { selectUser } from "@/pages/Auth/store/selectors";
+import { Logger, UserService } from "@/services";
+import { addMetadataToItemsBatch } from "@/services/utils";
 import {
-  ChatService,
-  CommonFeedService,
-  CommonService,
-  FeedItemFollowsService,
-  Logger,
-  UserService,
-} from "@/services";
-import { FirestoreDataSource, InboxItemType } from "@/shared/constants";
-import {
-  ChatChannelLayoutItem,
   checkIsFeedItemFollowLayoutItemWithFollowData,
-  FeedLayoutItemWithFollowData,
+  InboxItemsBatch as ItemsBatch,
 } from "@/shared/interfaces";
-import {
-  ChatChannel,
-  FeedItemFollow,
-  FeedItemFollowWithMetadata,
-  InboxItem,
-  Timestamp,
-} from "@/shared/models";
+import { InboxItem, Timestamp } from "@/shared/models";
 import {
   inboxActions,
   InboxItems,
@@ -34,164 +20,6 @@ interface Return
   fetch: () => void;
   refetch: () => void;
 }
-
-interface ItemBatch<T = InboxItem> {
-  item: T;
-  statuses: {
-    isAdded: boolean;
-    isRemoved: boolean;
-  };
-}
-
-type ItemsBatch = ItemBatch[];
-
-const addMetadataToFeedItemFollowItemsBatch = async (
-  batch: ItemBatch<FeedItemFollow>[],
-): Promise<ItemBatch<FeedLayoutItemWithFollowData>[]> => {
-  const data = await Promise.all(
-    batch.map(async ({ item }) => {
-      try {
-        const [common, feedItem] = await Promise.all([
-          CommonService.getCachedCommonById(item.commonId),
-          CommonFeedService.getCommonFeedItemById(
-            item.commonId,
-            item.feedItemId,
-          ),
-        ]);
-
-        return {
-          common,
-          feedItem,
-        };
-      } catch (error) {
-        return null;
-      }
-    }),
-  );
-  const parentCommons = await Promise.all(
-    data.map(async (item) => {
-      try {
-        return item?.common?.directParent?.commonId
-          ? await CommonService.getCachedCommonById(
-              item.common?.directParent?.commonId,
-            )
-          : null;
-      } catch (error) {
-        return null;
-      }
-    }),
-  );
-  const finalData = batch.map<ItemBatch<FeedItemFollowWithMetadata> | null>(
-    (batchItem) => {
-      const foundItem = data.find(
-        (item) => item?.feedItem?.id === batchItem.item.feedItemId,
-      );
-
-      if (!foundItem || !foundItem.common || !foundItem.feedItem) {
-        return null;
-      }
-
-      const foundParentCommon = foundItem.common.directParent?.commonId
-        ? parentCommons.find(
-            (parentCommon) =>
-              parentCommon?.id === foundItem.common?.directParent?.commonId,
-          )
-        : null;
-
-      return {
-        ...batchItem,
-        item: {
-          ...batchItem.item,
-          commonName: foundItem.common.name,
-          parentCommonName: foundParentCommon?.name,
-          commonAvatar: foundItem.common.image,
-          feedItem: foundItem.feedItem,
-        },
-      };
-    },
-  );
-
-  return finalData
-    .filter((item): item is ItemBatch<FeedItemFollowWithMetadata> =>
-      Boolean(item),
-    )
-    .map<ItemBatch<FeedLayoutItemWithFollowData>>((item) => ({
-      statuses: item.statuses,
-      item: {
-        type: InboxItemType.FeedItemFollow,
-        itemId: item.item.feedItemId,
-        feedItem: item.item.feedItem,
-        feedItemFollowWithMetadata: item.item,
-      },
-    }));
-};
-
-const addMetadataToChatChannelsBatch = (
-  batch: ItemBatch<ChatChannel>[],
-): ItemBatch<ChatChannelLayoutItem>[] =>
-  batch
-    .filter(({ item: chatChannel }) => chatChannel.messageCount > 0)
-    .map(({ item: chatChannel, statuses }) => ({
-      item: {
-        itemId: chatChannel.id,
-        type: InboxItemType.ChatChannel,
-        chatChannel: chatChannel,
-      },
-      statuses,
-    }));
-
-const addMetadataToItemsBatch = async (
-  userId: string,
-  batch: ItemsBatch,
-  feedItemIdsForNotListening: string[] = [],
-): Promise<ItemBatch<FeedLayoutItemWithFollowData>[]> => {
-  const batchWithFeedItemFollowItems = (
-    await Promise.all(
-      batch.map(async (batchItem) => {
-        const item =
-          batchItem.item.type === InboxItemType.FeedItemFollow
-            ? await FeedItemFollowsService.getFeedItemFollowDataById(
-                userId,
-                batchItem.item.itemId,
-                FirestoreDataSource.Cache,
-              )
-            : null;
-
-        return item ? { item, statuses: batchItem.statuses } : null;
-      }, []),
-    )
-  ).filter((item): item is ItemBatch<FeedItemFollow> => Boolean(item));
-  const batchWithChatChannelLayoutItems = (
-    await Promise.all(
-      batch.map(async (batchItem) => {
-        if (batchItem.item.type !== InboxItemType.ChatChannel) {
-          return null;
-        }
-
-        let item = await ChatService.getChatChannelById(
-          batchItem.item.itemId,
-          FirestoreDataSource.Cache,
-        );
-
-        if (item?.messageCount === 0) {
-          item = await ChatService.getChatChannelById(
-            batchItem.item.itemId,
-            FirestoreDataSource.Server,
-          );
-        }
-
-        return item ? { item, statuses: batchItem.statuses } : null;
-      }, []),
-    )
-  ).filter((item): item is ItemBatch<ChatChannel> => Boolean(item));
-  const batchWithFeedLayoutItemWithFollowItems =
-    await addMetadataToFeedItemFollowItemsBatch(batchWithFeedItemFollowItems);
-
-  return [
-    ...batchWithFeedLayoutItemWithFollowItems,
-    ...addMetadataToChatChannelsBatch(batchWithChatChannelLayoutItems),
-  ].filter(({ item }) => !feedItemIdsForNotListening.includes(item.itemId));
-};
 
 export const useInboxItems = (
   feedItemIdsForNotListening?: string[],
